backend/requirements.txt
fastapi[standard]==0.121.2
sqlmodel==0.0.27
psycopg==3.2.13
argon2-cffi==25.1.0